refactor: extract ThaiNamesGeneratorPage to tools feature directory
