+++ src/features/tools/pages/ThaiNamesGeneratorPage.tsx
@@ -1,0 +1,236 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Label } from "@/components/ui/label";
+import { Badge } from "@/components/ui/badge";
+import { useState, useRef } from "react";
+import {
+  Copy,
+  Check,
+  Shuffle,
+  Settings2,
+  User,
+  Building2,
+  Star,
+} from "lucide-react";
+
+import { generateThaiName } from "@/lib/data/thai-names";
+
+interface GeneratedName {
+  firstName: string;
+  surname: string;
+  nickname: string;
+}
+
+interface NameFieldProps {
+  label: string;
+  value: string;
+  icon: React.ReactNode;
+  isCopied: boolean;
+  onCopy: () => void;
+}
+
+function NameField({ label, value, icon, isCopied, onCopy }: NameFieldProps) {
+  return (
+    <div className="flex items-center justify-between gap-3 rounded-lg border bg-card p-3 transition-colors hover:bg-accent">
+      <div className="flex min-w-0 flex-1 items-center gap-3">
+        <div className="bg-primary/10 flex h-8 w-8 items-center justify-center rounded-md text-primary">
+          {icon}
+        </div>
+        <div className="min-w-0 flex-1">
+          <div className="text-xs text-muted-foreground">{label}</div>
+          <div className="truncate font-medium">{value}</div>
+        </div>
+      </div>
+      <Button
+        variant="ghost"
+        size="sm"
+        onClick={onCopy}
+        className="h-8 w-8 p-0"
+      >
+        {isCopied ? (
+          <Check className="h-4 w-4 text-green-500" />
+        ) : (
+          <Copy className="h-4 w-4" />
+        )}
+      </Button>
+    </div>
+  );
+}
+
+export function ThaiNamesGeneratorPage() {
+  const [count, setCount] = useState<number>(5);
+  const [generatedNames, setGeneratedNames] = useState<GeneratedName[]>([]);
+  const [copiedStates, setCopiedStates] = useState<Record<string, boolean>>({});
+  const resultsContainerRef = useRef<HTMLDivElement>(null);
+
+  const handleCopy = async (text: string, key: string) => {
+    try {
+      await navigator.clipboard.writeText(text);
+      setCopiedStates((prev) => ({ ...prev, [key]: true }));
+      setTimeout(() => {
+        setCopiedStates((prev) => ({ ...prev, [key]: false }));
+      }, 2000);
+    } catch (err) {
+      console.error("Failed to copy: ", err);
+    }
+  };
+
+  const handleGenerate = () => {
+    const names: GeneratedName[] = [];
+    for (let i = 0; i < count; i++) {
+      const gender = Math.random() < 0.5 ? "male" : "female";
+      const generatedName = generateThaiName({
+        includeFirstName: true,
+        includeSurname: true,
+        includeNickname: true,
+        gender: gender,
+      });
+
+      names.push({
+        firstName: generatedName.firstName || "",
+        surname: generatedName.surname || "",
+        nickname: generatedName.nickname || "",
+      });
+    }
+    setGeneratedNames(names);
+
+    // Scroll to results container after generating
+    setTimeout(() => {
+      resultsContainerRef.current?.scrollIntoView({
+        behavior: "smooth",
+        block: "start",
+      });
+    }, 100);
+  };
+
+  return (
+    <div className="container mx-auto max-w-6xl px-4 py-8">
+      <div className="space-y-6">
+        {/* Header */}
+        <div className="space-y-2">
+          <h1 className="text-3xl font-bold tracking-tight">
+            โปรแกรมสุ่มชื่อไทย
+          </h1>
+          <p className="text-muted-foreground">
+            สุ่มชื่อจริง นามสกุล และชื่อเล่นคนไทย
+            สำหรับทดสอบระบบหรือใช้เป็นข้อมูลตัวอย่าง
+          </p>
+        </div>
+
+        <div className="grid gap-6 lg:grid-cols-[350px_1fr]">
+          {/* Settings Panel */}
+          <Card>
+            <CardHeader>
+              <CardTitle className="flex items-center gap-2">
+                <Settings2 className="h-5 w-5" />
+                ตั้งค่า
+              </CardTitle>
+            </CardHeader>
+            <CardContent className="space-y-6">
+              {/* Count Selection */}
+              <div className="space-y-3">
+                <div className="flex items-center justify-between">
+                  <Label className="text-sm font-semibold">จำนวนชื่อ</Label>
+                  <Badge variant="secondary">{count}</Badge>
+                </div>
+                <input
+                  type="range"
+                  min="1"
+                  max="20"
+                  value={count}
+                  onChange={(e) => setCount(parseInt(e.target.value))}
+                  className="w-full"
+                />
+                <div className="flex justify-between text-xs text-muted-foreground">
+                  <span>1</span>
+                  <span>20</span>
+                </div>
+              </div>
+
+              <Button onClick={handleGenerate} className="w-full" size="lg">
+                <Shuffle className="mr-2 h-4 w-4" />
+                สุ่มชื่อใหม่
+              </Button>
+            </CardContent>
+          </Card>
+
+          {/* Results Panel */}
+          <div className="space-y-4" ref={resultsContainerRef}>
+            {generatedNames.length === 0 ? (
+              <Card>
+                <CardContent className="flex min-h-[400px] items-center justify-center p-8">
+                  <div className="text-center text-muted-foreground">
+                    <Shuffle className="mx-auto mb-4 h-12 w-12 opacity-50" />
+                    <p>กดปุ่ม &quot;สุ่มชื่อใหม่&quot; เพื่อเริ่มต้น</p>
+                  </div>
+                </CardContent>
+              </Card>
+            ) : (
+              <div className="space-y-4">
+                {generatedNames.map((name, index) => (
+                  <Card key={index}>
+                    <CardHeader>
+                      <CardTitle className="text-lg">
+                        ชื่อที่ {index + 1}
+                      </CardTitle>
+                    </CardHeader>
+                    <CardContent className="space-y-3">
+                      <NameField
+                        label="ชื่อจริง"
+                        value={name.firstName}
+                        icon={<User className="h-4 w-4" />}
+                        isCopied={copiedStates[`${index}-firstName`] || false}
+                        onCopy={() =>
+                          handleCopy(name.firstName, `${index}-firstName`)
+                        }
+                      />
+                      <NameField
+                        label="นามสกุล"
+                        value={name.surname}
+                        icon={<Building2 className="h-4 w-4" />}
+                        isCopied={copiedStates[`${index}-surname`] || false}
+                        onCopy={() =>
+                          handleCopy(name.surname, `${index}-surname`)
+                        }
+                      />
+                      <NameField
+                        label="ชื่อเล่น"
+                        value={name.nickname}
+                        icon={<Star className="h-4 w-4" />}
+                        isCopied={copiedStates[`${index}-nickname`] || false}
+                        onCopy={() =>
+                          handleCopy(name.nickname, `${index}-nickname`)
+                        }
+                      />
+                    </CardContent>
+                  </Card>
+                ))}
+              </div>
+            )}
+          </div>
+        </div>
+
+        {/* Info Section */}
+        <div className="space-y-2 rounded-lg border border-blue-500/50 bg-blue-500/10 p-6">
+          <h3 className="font-semibold text-blue-700 dark:text-blue-400">
+            ℹ️ เกี่ยวกับข้อมูล
+          </h3>
+          <p className="text-sm text-blue-700/80 dark:text-blue-400/80">
+            ฐานข้อมูล 22,000+ รายการจาก{" "}
+            <a
+              href="https://github.com/korkeatw/thai-names-corpus"
+              target="_blank"
+              rel="noopener noreferrer"
+              className="font-medium underline hover:text-blue-600 dark:hover:text-blue-300"
+            >
+              Thai Names Corpus
+            </a>{" "}
+            โดย Korkeat Wannapat (CC BY-SA 4.0)
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+}
